+++ changelog.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 * Don't switch to quoted-printable for long lines if already using base64
+* Fixed Travis-CI config when run on PHP 7
 
 ## Version 5.2.10 (May 4th 2015)
 * Add custom header getter